build using HCL2
* the hcl2template package can create a packer.Build from a set of HCL
(V2) files

had to make the already existing packer.coreBuild a public struct with
public fields

* Components interfaces get a new ConfigSpec Method.

This is a breaking change.

Current packer components: builder/provisioner/post-processor

each component interface now gets a `ConfigSpec() hcldec.ObjectSpec`
which allows packer to tell what is the layout of the hcl2 config meant
to configure that specific compoment.

This ObjectSpec is sent through the wire (RPC) and a `cty.Value` is sent
back through the already existing configuration entrypoints:

 Provisioner.Prepare(raws ...interface{}) error
 Builder.Prepare(raws ...interface{}) ([]string, error)
 PostProcessor.Configure(raws ...interface{}) error

"github.com/hashicorp/packer/helper/config".Decode is called from each
one of these configuration func and will configure the builder using any
found `cty.Value`.

The following was added:

 Communicator.Configure(...interface{}) ([]string, error)

Note that communicators are still part of a source but that we plan on
allowing a provisioner group to select a communicator in the future.
So the communicator block can be set in an HCL file but has no effect.